Made ChirpDBContext its own file with everything that felt like they would accompany properly
